requirements.txt
selenium>=4.16.0
webdriver-manager>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
